Switch all AST ptr-to-RCP to pure RCP
Issue # 53

This was the main goal of this exercise.  Now, all the AST nodes that
are allocated in the parser are allocated as RCPs, NOT pointers-to-RCPs.

Everything seems to pass.

I have not yet run this variant thru valgrind, but will soon.

I have also not yet removed the management of the "master" vector of
pointer-to-RCPs that is in the newExpression class.