Removes use of Teuchos RCP from Linear::MultiVector interface
To enable the creation of an abstract linear algebra interface for
multi-vectors that is also independent of Trilinos, tools like Teuchos::RCP
need to be removed from the method calls.  This commit changes the
getVectorView and getNonConstVectorView calls to return raw pointers
rather than Teuchos::RCPs.

Memory management of the returned raw pointer needs to be handled by
the calling classes.  Thus, all calls to this method needed to be
updated to delete the vector views.

Gitlab Issue #76