Add SON Bug 1318 to Xyce 7.2 Rel. Notes and Ref. Guide
This commit adds information on SON Bug 1318 to the Fixed Defects
Table of the Xyce 7.2 Release Notes, and to the .MEASURE section
of the Reference Guide.  That bug was about allowing for numerical
round-off issues if the FROM, TO and/or TD qualifiers on a .MEASURE
lines were expressions.

SON Bug 1318
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1318