Make configure/cmake use bison 3.0.4 min version
Issue #53

Variants became available in Bison starting with version 3.0.4, so it is
necessary to increase the required minimum from 2.4 to 3.0.4.

For all of our nightly test platforms, as well as our supported build
platforms, are all using a version of Bison that is >= 3.0.4.  So, this
should be safe to do.

Note, for the configure.ac file, it isn't set up to consider minor
version numbers.  So, technically, it is now set up to check for 3.0,
not 3.0.4.   In theory, a computer running 3.0.1 would pass configure,
but then not be able to compile the Bison input file.