Remove a ReactionParser warning
Because we had been continuing to support Bison 2.4 (for backward
compatibility to RHEL6), we have been ignoring all the warnings Bison
throws about "%name-prefix" in both the reaction and expression
parsers.

We have now upped our minimum supported Bison version to 3.0.4 so the
expression parser can use variants.  By the time of this release,
the deprecated %name-prefix had been replaced with "%define
api.prefix" instead.  Eric already changed the expression parser, I'm
now changing the reaction parser.

This removes one of the three warnings in the reaction parser.  There
is a second deprecated usage in there for api.parser.class, but that
can't be removed unless we make our minimum version of Bison be 3.3,
which we can't do because both RHEL7 and RHEL8 are stuck at 3.0.4.

The third warning out of the reaction parser is due to an extension of
the parser made to support cluster work, and bison is warning about an
unused element of a rule.  We should probably clear that up at some
point, too, but it is currently harmless.