Clear a bison warning from reaction parser
The presence of an empty action in a rule that needed no action was
causing Bison to spew a warning about an unused field in a parsed
line.  The field in question is a "specie" object.

Since the empty action (a comment inside braces) was there solely as a
place holder for an extension that has not happened since the code was
written, but the rule is used simply to recognize "SPECIE;" as a valid
"enhancment", putting a "delete $1" in the action is wrong and causes
a crash later in the one test that uses this feature.  However, having
NO action at all clears the warning and does not cause a crash.

I have removed the empty action to shut up the bison warning.

If the cluster work is ever revisited and the extension for which a
place was being held ever actually implemented, a real action that
makes use of "$1" can be inserted again.

This leaves only one warning from the reaction parser, one that cannot
be cleared until we are able to change the minimum acceptable Bison
version to 3.3 (maybe years from now, when RHEL gets with the pogrom
and updates the bison they include in their default package repos).