Allow Bison 3.7 for Xyce builds
The autotools build not only tests that Bison is new enough to work
with Xyce, but also tests that it is a version of Bison we've actually
tested.  It does this by having an explicitly enumerated list of
acceptable versions.  There's a reason for this, and it's been
discussed too much.

Bison 3.7 is out and works just fine (according to Rich).  This is the
default version of Bison on the latest OS X version anyway.

I have added bison 3.7 to the list of acceptable versions in
configure.ac.  We can expect external users of Xyce 7.2 to start
encountering bison 3.7 anyway, and this at least prevents the
inevitable questions on the forum.

This was discussed in the comments of issue #53, though it is not
actually directly a part of that issue at all.