Change "emplace" to "build for 3.0.4 compatibility
My recent changes to use variants in the expression parser
broke many test builds, disappointingly.

The reason is that while variants started being supported in version
3.0.4 (our new minimum version), the usage of variants changed in
version 3.2.

Prior to 3.2, the proper way to invoke a variant was using the "build"
function.  Starting with 3.2, they changed "build" to "emplace".

Fortunately, it appears that they preserved backward compatibility.  If
I change all the instances of "emplace" with "build" in the
ExpressionLexer.l file, the code still builds under Bison 3.5.

At some point they will probably drop this particular backward
compatibility, so we need to watch out for that.   Hopefully they have
not dropped it as of version 3.7.  We'll find out soon when Rich tries
this change out.