Code Cleanup and Fix in Measure::Base Functions
This commit moves the code, used within the measure classes, for
determining if two real numbers are "equal" to its own function.
This should improve maintainibility.

This commit also fixes an omission in the Base::isInvalidTimeWindow()
function.  The previous version did not account for the TD qualifier
that can be used by TRAN measures.

SON Bugs 1304 and 1318
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1318