Add tests for sqrt(-neg) and log10(-neg)
This use case, which for real numbers will result in Nan normally,
needed to be tested.  In Xyce, these have to return usable numbers.

For the real case, the return 1e50.

For the complex case, they just return the complex result.  But, b/c
these can be opposite sign, in the complex case I compare the abs value
of the results.