Remove minimum time-step restriction for TRA device
The lossless transmission line works by keeping a history of accepted
time, voltage pairs at each end of the line for every time step that
succeeds.  In order to calculate the currents at each end at time T, it
needs to determine what the voltage at the other end was at time T-TD,
where TD is the line delay.

It gets this time-delayed value by using the history to interpolate ---
odds are very good that the time integrator did not actually evaluate
the solution at T-TD.  It uses a Lagrange interpolation formula, and
tries to find three values T1, T2, and T3 such that T1 > T-TD > T2 >T3.
This is done simply by searching for T1 in the history, and then taking
the two previous times.  The history is preloaded with three values that
duplicate the DC solution at 0, -TD and -2TD.

When this device was first written in 2001, I tried to copy what it did
from what the SPICE3F5 device was doing.  And at the time I noticed that
when the time step was greater than TD the interpolation broke, and so I
inserted both a test to make sure we weren't extrapolating *and* code to
force the time integrator never to take a time step larger than TD.

This is NOT what SPICE3F5 does, but I was unable to figure out exactly
what SPICE really *did* do, as it looks like the method it uses should
fail horribly when the simulator takes a step larger than the line
delay, because it seems to be assuming it can always find a value in the
history that is *later* than T-TD, and looks like it should then be
accessing invalid data if it can't find such a time.  This misread of
the SPICE3F5 code led to my writing the time-step restriction in.

However, there is no reason that the Lagrange formulae can't be used to
extrapolate, and if the time integrator is allowed to pick its own time
step irrespective of the line delay, all we need to do is to check if we
can find T1 > T-TD, and if we *CAN'T* than use T1, T2, and T3 as the
last three elements of the history.  Extrapolation will happen
naturally.

This commit changes the InterpV1V2FromHistory method so that if it
cannot find a time in history that is later than T-TD it simply uses the
last element of the history and the two prior as the Lagrange
interpolation points.  It also makes the "maxTimeStepSupported" method
return "false" instead of "true", so that "getMaxTimeStepSize" is never
called.

In my testing, this has dramatically improved performance of some
circuits with very small transmission line delays that were
unnecessarily having their timesteps restricted to something much
smaller than the step error control wanted to use.

The transmission line does still look at the slopes of the input signals
looking for discontinuities.  If it finds what it thinks is a
discontinuity, it still sets a breakpoint at T+TD so that the time
integrator catches the discontinuity at the other end when it gets
there.  This *IS* what SPICE3F5 does and what we have always done as
well.  It's only the minimum timestep that was a uniquely Xyce mistake.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/33