Connect random expression operators to SAMPLING
Issue # 84

Also, on the SON bugzilla this is bug 1058.

As part of my work on the new expression library, I had been working on this,
and was very close to getting it working in July 2020.  But, by then I was
getting concerned that the new expression library work had dragged on too
long, and as this issue represented a completely new feature, I decided
that it should be a lower priority than getting the first cut of the
library done.   So, at that point, I suspended that work and focused on
reproducing the behavior of the old library vis-à-vis random operators
such as AGAUSS.

Anyway, as the new expression library is now merged into the development branch,
and seems to be working, it seemed like a good time to revive this issue
yesterday.

As it turns out, I was VERY close to getting a preliminary version running.
It only took about a hour of coding to make it work for .SAMPLING.

This isn't finished, but it is pretty close, at least in terms of the
code.

Things that still need to be done include:

  - It probably doesn't work in parallel.  I am aware of one place in
    the code that needs fixing for parallel.

  - I have only set it up for .SAMPLING, and not .EMBEDDEDSAMPLING or
    .PCE.  But that won't be very hard to do.

  - I haven't tested it very much yet, or created any regression tests.
    So, possibly some problems will emerge when I start doing this.

  - I haven't made SAMPLING be truly parallel, in the "embarassingly
    parallel" sense, but that will have to wait until after 7.2.

Most of the code involved in this capability is pretty solid.  ie, the
UQ functions that I implemented a few years ago to support sampling,
etc, and the new expression library itself.  The main new thing added by
this commit is glue connecting them together.  So, I have some
confidence that this can be made production soon.  But we'll see.