Fix to make giant tables fast w/variants
Issue #53

I recently implemented variants in the Bison parser from the new
expression library.

The advantage of doing this (I hoped) was that it would allow me stop
using pointers-to-RCPs in the parser.  And, if I could get rid of the
parser-to-RCPs, then I could also get rid of a data structure in the
newExpression class, which doubled the memory required to store the AST
tree.  And, finally, getting rid of this allowed the removal of the
newExpression destructor as well.

This worked, but I was too overzealous with the removal of pointers.  In
addition to removing pointers to RCPs, I also removed a number of
pointers to std::string and also to std::vector<RCP>, etc.  This mostly
didn't seem to cause problems at first.

But I discovered today that it caused at least one significant performance
problem.  Large tables (like 50,000 entries, etc) are parsed as
std::vector<RCP<astNode> > and they slowed down like crazy.  They slowed down
because these large std::vector had to be copied every time a new
elemented was added to it.  So, it was copied 50,000 times, and each
time it was 1 element larger.

I had two test circuits that I was running manually.  Pre-variants, both
of these circuits took about 5 seconds to run.  Post-variants, they took
many minutes (I killed the run b4 I knew exactly how many). In one test
the table had about 40,000 entries and in the other it had 100k.

After verifying that this was the problem, I decided to try and find a
middle ground.  I really wanted to keep using variants, and that seemed
to be fine, in that I could keep using pointers for everything, and use
variants.  But, I also wanted to get rid of the newExpression destructor
and the superfluous memory usage.

It turns out that this was possible to do.  I just had to make sure
that all argument lists of arbitrary length (such as tables) are still
passed around as pointers to std::vectors, not as std::vectors.  While I
was at it, I went back to having everything that wasn't part of the AST
tree still use raw pointers.     The AST tree itself is still completely
based on RCPs, without raw pointers.

So, with this commmit, the performance of giant tables is restored, but
I still am using variants, I am no longer wasting memory, and th
newExpression destructor is still the default constructor.  The two test
cases are now back to running in about 5 seconds.