Remove long-obsolete text from error message
Back in 2004, we discovered that restarting netlists that had
transmission lines in them caused segfaults.  That was because the
original version of the transmission line was written before the restart
capability, and no code was put into the TRA to make sure to store its
history in restart files (or reload that history when restarting).  As a
result, the code tried to interpolate using an empty history and puked.

The first "fix" for this was not a fix at all --- it simply replaced the
hard crash with a fatal error and comprehensible error message saying
"Restarts of netlists with transmission lines does[sic] not work yet".
That was in commit d8248394 in June 2004.  A real fix was then put in to
save and restore history to/from the restart file in commit 07b1bbfc in
Sept of 2004.

When the fix was really complete, the error message text was not
changed.

It is still an error to try to interpolate from an empty history list,
but now it is wrong to tell the user that the empty history list might
have come from trying to restart.

This is almost irrelevant to anything, as it is now impossible for this
error condition to come up.  But if it ever does come up during
rebugging, the error message should not cast aspersions in the wrong
direction.

https://morannon.sandia.gov/bugzilla/show_bug.cgi?id=442