Improves '.options linsol' documentation
It was noted in SON bug #1290 that the documentation of the options
provided by the linear solver package through '.options linsol' was
missing some commonly recommended options.  The documentation of the
reference guide and user guide has been reviewed and missing options
have been added to the tables and discussion in both documents.

Bug:  1290 SON