Fix Valgrind Error in DAC Device
This commit fixes a memory error in the Instance::updateTVVEC()
function in the DAC device.  This is probably not the best fix;
that fix would require a complete re-write of updateTVVEC() to:

  1) remove some previous hacks done for (the now defunct) Habinero
code, at t=0.

  2) more properly handle invalid pairs of invalid time-voltage
pairs.  Examples include an empty set of pairs, and a set of pairs
where all of the time values are less then the current accepted
time-step.

SON Bugs 1052, 1062 and 1067
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1052
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1062
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1067