Fix expression-based sampling for parallel
Issue #84

Over the weekend a created a handful of tests for expression-based
(AGUASS, etc) random sampling and embedded sampling.  Unfortunately,
some of these failed in parallel during regression testing.

There were two reasons for the failures.

The most important reason these tests were failing in parallel had to do with
gathering up all the expression-based random operators so that the SAMPLING and
other UQ methods would have a master list of random inputs in one place.
I previously created a function in the device manager to do this, called
DeviceMgr::getRandomParams.

Unfortunately DeviceMgr::getRandomParams was broken in parallel, in the
second half of the first.

The first part of that function worked fine in parallel.  In that part, it
would first gather up expressions containing random operators from the list
of global parameters.  As this list is a global list, it is known on
every processor.  So, this worked fine.

However, in the second part of the function atttempted to gather random
operators from device parameters (models and instances).  These are
not generally known on every processor.  And, the data structures I was
using are purely local.  So, for tests that applied AGAUSS to a device
parameter, such as R1:R, the list of operators was different on each
processor and this caused everything to break.

This issue didn't have a workaround at all, so I had to write some code
to fix it.    The DeviceMgr::getRandomParams function populates a
std::vector of "Analysis::SweepParam" objects.  So, I created
packedByteCount, pack, and unpack functions for this object, and I also
added code to the DeviceMgr::getRandomParams function to apply these
when running in parallel.  This seems to have fixed the problem.

The other, less important reason for these tests failing had to do with
setting the random seed from the mainExpressionGroup.  This is ONLY set
when the code is using the old expression behavior when -hspice-ext all
is not enabled.

I made the mistake of using the Analysis::UQSupport::getTheSeed function  (which
I wrote a while ago to support sampling) to get the seed into the group class.
But, this "getTheSeed" function assumes it is called on all processors, and
in parallel, coming from an expression group, this was not the case.   This
had the workaround of the command line option -hspice-ext all, so it wasn't as
catastrophic as the other issue, above.  But, as long as we're still supporting
the old use case, it needs to work correctly in parallel.  So, I have fixed
that as well.