Fix device-level params for random expr operators
Issue #84

My previous work on this issue handled global parameters correctly, but
device level (instance and model) parameters incorrectly.

The main function in question for this commit is the
DeviceMgr::setParameterRandomExpressionTerms function.  It was a copy of
the much older function, DeviceMgr::setParameter, plus modifications.
As it turns out, I had only modified the part that pertains to global
parameters, and left the device params part of the function almost
completely alone.

This was a mistake, but it was well  hidden by the
fact that most of my the sampling/embeddedsampling tests are pretty
simple.  If you have this in the netlist:

R1 1 0 {agauss(3k,1k,1)}

then the function worked just fine.  But, if you had this instead:

R1 1 0 {3.0*agauss(3k,1k,1) + rand()}

(in other words, a more complicated expression for R1:R, that just
happened to include random operators) then the function did the wrong
thing.

The intent of the function is to provide values from the sampling
algorithm for each random operator.  It is not intended to provide
parmaeter values directly.   So, for example, the function would
pass in the value for the "agauss" operator.   It did not pass in the
value of R1:R, as figuring that out is the job of the expression
library, not the sampling algorithm.  As the device-level param code as
copied from "setParameter" and unmodified, it was treating the value of
"agauss" as the value for R1:R.  For most of the certification tests,
this is correct.  But it is incorrect if you try to have a more
complicated expression.

So, I have updated the code in that part of the function, and added a
new code to the device entity class to support it.  Now, the function
will, for device-params (1) find the correct device enty, (2) find the
dependent param, (3) access the expression of that depedent param and
(4) set the value of the random operator inside of that expression (5)
re-evaluate the expression (6) call the relevant processParam functions.

Note, if a parameter expression has multiple random operators in it,
this code should handle it fine.  Each random operator within an
expression has a unique index, which is invoked on this function call.

This issue was preventing the embedded_cmosInverter circuit from working
correctly with expression-based random operators.