Add SON Bug 1309 to Known Defects in 7.2 Rel. Notes
This commit adds SON Bug 1309 to the Known Defects table in the
Xyce 7.2 Release Notes.  That bug is about issues with the FROM and
TO qualifiers not being equal to an accepted time-step (or a requested
AC, DC or NOISE sweep value) for the AVG, INTEG and RMS measures for
all measure modes.

SON Bug 1309
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1309