Add "exp" as an allowe constant in expressions
I discovered this as I started working on documentation for the new
expression library.  I am not sure how much needs to be added to the
users manual, as the new library mostly reproduces the old one.  But, to
check, I started reading the ABM chapter.  Early in that chapter, it
mentions that "EXP" is an allowed constant.  This suprised me, as I had
not implemented it.

As it turns out, this is a feature that was supported in the old
expression library.  where "exp" by itself, without any parens or argument
following it, is supposed to be a constant = exp(1.0).  Apparently, this
use case is not covered by any existing regression test, something that I
will be fixing shortly.

This commit thus adds "exp" as an allowed constant=exp(1.0).  The lexer now
supports a new token: TOK_EXP_CONST, and this token is used in the parser
to allocate an expOp class with the argument of 1.0.

This commit also creates 3 new unit tests for this feature.  They pass.