Make second seed for resampling conditional
For sampling and embedded sampling, it is possible to do polynomial
chaos approximations on the outputs.  When this is done, an
polynomial-based model of output uncertainties is created.  Once this
polynomial-based model has been set up, it is possible to perform LHS
sampling on that polynomial.

This can be useful for various reasons, and is very, very cheap to do,
as the sampling routine just has to re-evaluate the polynomial
approximation once per sample point.

So, back when I was working on this stuff a few years ago, I added that
feature.  It is turned on by a "resamplePCE" boolean parameter.

For some reason when I did this, I made it unconditionally create a new
random number seed for the resampling algorithm.  And, for the case of
.SAMPLING, I made it re-generate this seed if ANY outputs were
requested, not just resampling outputs.  The function that generates
the seed, by default, outputs the seed to std output, so this meant that
in some cases I was getting 2 seeds printed to std output.

This commit doesn't 100% get rid of the issue, but it at least puts the
generation of a new seed behind the "resamplePCE_" if statement.  So, it
will only print this seed to the screen if resampling has been
requested, which is not the usual case.

This does beg the question of whether or not a second seed makes any
sense.

It does make sense to generate a "resampling" seed when applying sampling
to a PCE approximation that was generated via quadrature. The PCE
algorithm isn't using sample points in that case, so the resampling seed
is the first seed generated.    So in that case it is necessary.

But for regression PCE, which uses sampling points as inputs to the
regression algorithm in the first place, it doesn't seem necessary
to generate a new separate "resampling" seed.  It isn't clear to me
why one would bother.  So, probably I'll change that later.

But for now, the if-statement is enough to get rid of 99% of these
unnecessary "seed" outputs.