Fix so that epxression random ops work w/PCE
Issue #84

This fixes a memory bug that was causing a seg fault, if I tried to
apply PCE methods to a sampling run that used expression-based random
operators such as agauss.

99% of the code was in place for the non-intrusive forms of PCE to
work with the expression-based random operators, as they are post-process
steps taken after sampling.  And, of course, sampling is set up to work
this those operators now.

However, there is one data structure (a vector of parameter names) that
I was using to partially intialize the various stokhos objects, and in
the new code this object was not getting set up.  This led to a memory
error in the PCE setup.   Fixed now.