Remove a sometimes-incorrect warning.
The ExpressionData::setup function was significantly rewritten for the
new expression library.  I've noticed that in parts of the code where I
had to do this, occasionally something got lost in translation.

The setup function purpose is to to fully resolve the expression that is
wrapped by the ExpressionData class.  ExpressionData is mostly for
wrapping .print line expressions.  So, the setup function will attach
external functions, external global parameters, etc.

It also throw out a warning if it isn't able to resolve something from
the expression, to the effect of "This field is not resolvable".

I noticed today as I was working on tests that this warning would
occasionally get issued, and it would be wrong.  For example, in a
circuit I was working on, I had {R1:R} on the .print line, which was
perfectly fine.  But the warning got issued, saying that R1:R was not
resovable.

This was due to the "setup" function incorrectly assuming that if an
unresolved string was not in the .param or .global_param containers,
it must be unresolvable.  This wasn't correct.  R1:R isn't a .param
or a .global_param but it is a resolvable entity.  It just doesn't
need to be resolved here in ExpressionData::setup.

The "outputs group" is really what ultimately has to
resolve a device parameter like R1:R, and it does so just fine.  And, if
R1 is not present in the circuit, and is thus NOT resolvable, the
outputs group issues a fatal error early in the calculation.

So, this commit comments out the warning from the ExpressionData::setup
function, as it is redundant and sometimes wrong.