Update to indicate FREQ allowed in ABM sources
I have started to work on documenting the new expression library.

Expressions are discussed in various places in the User and Reference
guides.  As time is short, and 99% of the new expression library work
was to reproduce the old capability, I am hoping that most of the
documentation can remain as-is, with just occasional fixes for things
that are now obsolete.

This commit fixes one such issue.  The text of the ABM chapter stated
that "FREQ cannot be used in ABM sources".

I was pretty sure, based on my expression work, that this wasn't true
any more.  So, I wrote a little test program, and it looks like I was
correct about that.  So I have updated the text to reflect this.