Fix my "fix" for parallel random expression ops
Issue #84

Earlier this week I fixed the code that assembles a global master vector
of random expression operators in the device package.  Ostensibly, the
fix was so that this code would work correctly in parallel.  It seemed
to work at the time.

However, today I discovered that I was wrong about that.  The tests I
had used to ensure my previous fix worked were too simple.  As such,
I had two problems.

(1) my fix from earlier in the week had an indexing mistake
(2) I was incorrectly assuming that model params were unique to the
local process, like instance params are.  This is clearly wrong.

Both of these issues are now fixed.  My fix is a bit clumsy, but it
works.  This had better be the final fix for this, as I am out of
time, in terms of the 7.2 release.