Fixing outdated text on expressions in user guide
Issue #38, new expression library documentation.
Also, bug 1249 on the SON bugzilla.

I have been working on documenting the new expression library.  As it
replaces a mature feature, there is already a lot of documentation that
is roughly correct.  However it has been a long time since any of this
was gone over carefully, I think.   So, I am primarily correcting
out-of-date text for this issue.

Earlier I changed text that claimed that FREQ was not allowed to be used
in ABM devices.  That isn't true anymore, and I created a regression
test to prove it.  I found another place (in the chapter 4 on params and
expressions) which also made this claim, and am correcting that with
this commit.

Also, I found this language in the ABM chapter, which is clearly wrong:

------------
Parameters defined within a given scope can be used in any expression within that scope. The only limitation on ordering is for the use of a parameter in an expression that defines the value of another parameter. In that case, all parameters used in the expression must be defined before being used to define another parameter. So, in the following example:

  \begin{vquote}
R1  1  0  \{B+C\} ; OK because the expression is not used to define a param
.PARAM A=3
.PARAM B=\{A+1\}  ; OK because A is defined above
.PARAM D=\{C+2\}  ; Illegal because C is not yet known
.PARAM C=2
\end{vquote}
------------

I am not sure how this particular bit of text managed to remain in the
user guide for so long.  I don't think this has been correct for many
years.  Anyway, it is removed now, as of this commit.