Updates for new expressions
Issue #38

This commit updates the expression section of the reference guide to be
consistent with the new expression library.

This includes adding the following new operators to the table:  Re(),
R(), Img(), Ph(), AUNIF, and UNIF.

It also modifies the text about complex arithmetic in the expression
library.  Basically, at this point it can be done without having to
manually specify a real equivalent form in the expression.  But it still
by default returns only the real part on the .PRINT line.  So, if the
user wants more than that (like the imaginary part), they must specify
it via {Img(x)}, or something like that.