Added a section about sampling w/random operators
Issue #84
Issue #38

This adds a very small amount of documentation about linking the random
expression operators, such as AGAUSS to SAMPLING.  It is just enough for
the reader to know it exists but not much more than that.

I consider this feature to be very much a "beta" feature, so I think
that is enough for now.