Remove conditionals that result in a C99 issue
Not surprisingly, ksparse has come code that does not comply with the
C99 standard. On macOS 10.15 (really Xcode 12), compliance with C99 has
become more rigid. As a result, Xyce fails to build on Catalina without
this change. Basically, the code is now unconditionally including
`stdlib.h`, and not doing its own thing.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/69