Discontinue bundling of documentation in installers
The CMake build (currently used only on Windows) no longer bundles
documentation with the installer.  Nobody noticed this until last night.
All the CMake build is doing is dumping README.TXT into the doc
directory.  README.TXT is useless.

It has been discussed that this would have been the right thing to do
on other platforms anyway, because inclusion of the documentation in
the installers makes the installers dependent on R&A of the docs,
slowing down QA.

This commit changes the make_release so it simply ignores all the
command line arguments that relate to documentation, and doesn't
bundle any of them.

It also changes README.TXT so it simply tells the user where to find
the documentation on our web site.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/90