Small Re-factoring of Functions for .PRINT Wildcards
This commit does a small re-factoring of the functions used for
handling wildcards on .PRINT lines.  The removeStarVariables()
function was re-named, since it now handles both * and ? wildcards.
In addition, some common code related to handling Y devices that
should not be included in the wildcard-specified variable lists has
been moved to its own function.  This commit also adds declarations
of those two functions to the associated .h file.

SON Bug 1319
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1319