Add more doc for use of AGAUSS/etc with sampling
This is for bug 1058 SON

Also issue #84

Bug 1058 SON has been closed, and mostly documented.  However, I
neglected on aspect of documentation, which was the sampling and
embedded sampling sections of the reference guide.

This was pointed out by Tom.

This commit adds some text about how to use random operators with
sampling and embedded sampling to their sections of the reference guide.
It also adds a note to the "differences with other simulators" section,
to explain that it doesn't hand the "local vs global" variation issue in
the same way as Hspice.  (ie they handle it and we do not)