Implement spline function in expression library
Issue #80

I started working on this a while ago, but did not finish it in time for
the 7.2 release.

As it turns out, I was only about 1/2 away from getting it done.

This is very similar to "TABLE", but does not replace it.  It can be
used as an alternative, as it supports all the same formats that TABLE
does.  The only difference is that they keyword is SPLINE.

Since table supports a bunch of formats, it has a bunch of rules in the
lexer and parser.  As such, I had to create new versions of all of these
for SPLINE.   That feels kind of bloated but for now it works.

The TABLE capability used the tableOp class.  This new SPLINE capability
uses the interpolatorOp class, which is nearly the same, except that  is
uses the akima spline interpolator, rather than the linear interpolator
used by TABLE.

There are several other interpolators I could have used, including
wodicka splines, cubic splines and barycentric lagrange interpolation.
Possibly I will expand this capability later so that all of these will
be options that the user can choose.  But for now, Akima is it.

I also copied all the "TABLE" tests in the parserUnitTests.C program,
and created equivalent SPLINE tests.  All but one passed right out of
the box.  (which suggests that at least for these tests, there isn't
much difference between TABLE and SPLINE).  For the one that failed, I
was able to get it to pass by switching from EXPECT_EQ to
EXPECT_FLOAT_EQ, which has a tolerance for floating point comparisons.