Add Info on ? Print Wildcard to Reference Guide
This commit adds info on the ? (meaning any one character) wildcard
to the .PRINT section of the Reference Guide.

It also updates the section on legal device/node names to include
info on the ? wildcard.  A caveat about using ternary operators
(? and :) in node/device names was also added to that section.

SON Bug 1319
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1319