Add support for TEMPER/HERTZ, synonym for TEMP/FREQ
Issue #93

The TEMPER part of this is from Bug 50 SON:
"Hspice parameter TEMPER causes "parameter/function resolution issues""

This is a very old HSPICE-support bug.

It took about 5 minutes, and just required that I add "temper" and "hertz"
to the lexer file.

This commit also adds 2 TEMPER unit tests, which are identical to the 2 unit
tests for TEMP.

This commit also adds 2 HERTZ unit tests, which are identical to the 2 unit
tests for FREQ.